+++ src/models/mod.rs
@@ -1,3 +1,4 @@
 pub mod article;
 pub mod feed;
+pub mod parsed_feed;
 pub mod persisted_config;
