+++ src/providers/feed_parser/feed_parser_trait.rs
@@ -1,24 +1,6 @@
+use crate::models::parsed_feed::ParsedFeed;
+
 use super::Result;
-use chrono::{DateTime, Utc};
-
-/// Common structure representing parsed feed data
-#[derive(Debug, Clone)]
-pub struct ParsedFeed {
-    pub title: String,
-    pub link: String,
-    pub items: Vec<ParsedItem>,
-}
-
-/// Common structure representing a parsed feed item
-#[derive(Debug, Clone)]
-pub struct ParsedItem {
-    pub title: String,
-    pub link: Option<String>,
-    pub guid: Option<String>,
-    pub content: Option<String>,
-    pub author: Option<String>,
-    pub pub_date: Option<DateTime<Utc>>,
-}
 
 /// Trait defining the common interface for all feed parsers
 pub trait FeedParser: Send + Sync {
